Item -component ja QuestionList2 -component muokattu ja korjattu navigaatio. Listanäkymässä lisätty dynaaminen lista oikeaan palstaan.
