+++ src/components/Home.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import Item from './QuestionList';
+import Item from './Item';
 export default class Home extends Component {
     render() {
         return (
