+++ src/components/QuestionList.jsx
@@ -3,16 +3,18 @@ import QuestionDetails from './QuestionDetails'
 
 
 export default class QuestionList extends Component {
-    
+
     render() {
         const questionitems = this.props.items.map(item => {
-                return (
-                    <QuestionDetails question={item} key={item.id} />
-                )
-            })
+            return (
+                <QuestionDetails question={item} key={item.id} />
+            )
+        })
         return (
             <div className="QuestionList AppComponent">
                 <h2>QuestionList</h2>
                 {questionitems}
             </div>
-        )
+        );
+    }
+}
